requirements.txt
tensorflow
numpy
pandas
matplotlib
pandas_datareader
Pillow
scikit_learn
